requirements.txt
Pillow
python-docx
keyboard
pywin32
